refactor(Pieces): 게임 계산 로직 이전
