+++ src/main/java/chess/domain/piece/Pieces.java
@@ -7,6 +7,11 @@ import chess.domain.position.Row;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+import static java.util.stream.Collectors.counting;
+import static java.util.stream.Collectors.groupingBy;
 
 public class Pieces {
     private final List<Piece> pieces = new ArrayList<>();
@@ -68,5 +73,43 @@ public class Pieces {
 
     public void delete(Piece piece) {
         pieces.remove(piece);
+    }
+
+    public double score(Color color) {
+        return calculateGeneralScore(color) + calculatePawnScore(color);
+    }
+
+    private double calculateGeneralScore(Color color) {
+        return pieces.stream()
+                     .filter(piece -> piece.isSameColor(color))
+                     .filter(piece -> !piece.isPawn())
+                     .mapToDouble(Piece::score)
+                     .sum();
+    }
+
+    private double calculatePawnScore(Color color) {
+        return pawnCountingByColumn(color).values()
+                                          .stream()
+                                          .mapToDouble(this::lowerPawnScore)
+                                          .sum();
+    }
+
+    private Map<Column, Long> pawnCountingByColumn(Color color) {
+        return pawnsWith(color).stream()
+                               .collect(groupingBy(Piece::getColumn, counting()));
+    }
+
+    private List<Piece> pawnsWith(Color color) {
+        return pieces.stream()
+                     .filter(piece -> piece.isSameColor(color))
+                     .filter(Piece::isPawn)
+                     .collect(Collectors.toList());
+    }
+
+    private double lowerPawnScore(long count) {
+        if (count >= 2) {
+            return count * 0.5;
+        }
+        return count;
     }
 }
